Move partially qualified registration methods to extension class
